% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sequences.R
\name{extract_sequences}
\alias{extract_sequences}
\title{Write sequences associated with sampled IDs to a multifasta file.}
\usage{
extract_sequences(taxlist, seq_file, verbose = TRUE)
}
\arguments{
\item{taxlist}{list containing vector `$outputIDs` of sampled IDs (returned
by [run_TS()]).}

\item{seq_file}{character string with the path to the multifasta file
containing the input sequences.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
Updated `taxlist` containing field `outputSeqs`, a data frame with
information on the sampled sequences.
}
\description{
Write sequences associated with sampled IDs to a multifasta file.
}
