% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxlevel.R
\name{extract_taxlevel}
\alias{extract_taxlevel}
\title{Extract the desired ancestry level of a list of taxonomy IDs}
\usage{
extract_taxlevel(taxIDs, taxlevel, nodes)
}
\arguments{
\item{taxIDs}{vector of taxonomy IDs of interest}

\item{taxlevel}{desired taxonomic level to be extracted (e.g., `"order"`)}

\item{nodes}{data.frame containing the pre-processed information about
the NCBI taxonomy structure. This can be generated, e.g., by using
[get_counts()], or as object `$nodes` of a `taxonsampling` object.}
}
\value{
Data frame containing the taxonomy IDs of interest and a column with
the corresponding taxonomy IDs of their desired ancestry level.
}
\description{
Extract the desired ancestry level of a list of taxonomy IDs
}
