% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Return the relevant counts for running TaxonSampling}
\usage{
get_counts(
  taxonomy_path = NULL,
  ids_file = NULL,
  ids_df = NULL,
  spp_file = NULL,
  spp_df = NULL,
  nodes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{taxonomy_path}{path to folder containing the NCBI taxonomy files
(i.e., the extracted contents of _taxdump.zip_, which can be downloaded from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/> or retrieved using
[retrieve_NCBI_taxonomy()]).}

\item{ids_file}{path to a tab-separated file containng two two columns,
with the input taxon IDs in the first
column, and the corresponding sequence IDs (corresponding
to the ID strings int the multifasta input file, without
the ">" line starter) in the second column.}

\item{ids_df}{two-column data frame with the input taxon IDs in the first
column, and the corresponding sequence IDs (corresponding
to the ID strings int the multifasta input file, without
the ">" line starter) in the second column. Ignored if
`ids_file` is not `NULL`.}

\item{spp_file}{path to a tab-separated file containng two two columns,
with the input taxon IDs in the first column, and the corresponding number of
known species in the second column. If both `spp_file` and `spp_df` are `NULL`, then
the counts are computed internally based either on `nodes` or the
taxonomy files under `taxonomy_path`.}

\item{spp_df}{two-column data frame with the input taxon IDs in the first
column, and the corresponding number of known species in the second column.
This can be generated using [get_taxID_spp_counts()].
Ignored if `spp_file` is not `NULL`.}

\item{nodes}{data.frame containing the pre-processed information about
the NCBI taxonomy structure. This is generated either by using
[CHNOSZ::getnodes()], or as a result of a previous call to
[get_taxonomy_counts()]. If `nodes` is not `NULL` then `taxonomy_path` is
ignored.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
list object of class _taxonsampling_, containing:
\itemize{
    \item `$ids_df`: data.frame with taxon IDs in column 1 and corresponding
    sequence IDs in column 2, as loaded from `ids_file` or passed directly as
    input. Filtered to maintain only IDs that exist in `$nodes` and to remove
    duplicated IDs.
    \item `$nodes`: data.frame containing the pre-processed information about
    the NCBI taxonomy structure, extracted from file _nodes.dmp_ of the
    taxonomy files. Filtered to keep only nodes with IDs present in `$ids_df`
    and with a non-zero total count.
    \item `$countIDs`: numeric vector with the counts of the number of
    taxonomy nodes (of all levels) under each taxon ID.
    \item `spp_df`: two-column data frame with the input taxon IDs in the
    first column, and the corresponding number of known species in the second
   column. Filtered to have only the IDs present in `$countIDs`.
   \item `$countSpp`, a numeric vector with the counts of (known) species for each
   taxon from `taxlist$countIDs`.
}
}
\description{
This function evaluates the taxonomical structure from a set of input IDs
and returns the count of each taxon ID with non-zero occurrences in the
taxonomy graph, as well as the count of known species under each taxon.
}
