% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_df.R
\name{get_summary_df}
\alias{get_summary_df}
\title{Generate a summary dataframe from the result of run_TS}
\usage{
get_summary_df(taxlist)
}
\arguments{
\item{taxlist}{list object of class _taxonsampling_, returned by
[run_TS()]}
}
\value{
Data frame containing fields `$taxID`, `$level` (taxonomic level),
`$name` (scientific name), `species_count` (total number of species or leaf
nodes under that taxonomy ID), `$sample_count` (number of sampled species or
leaf nodes under that taxonomy ID) and `$seq_count` (number of available
sequences under that taxonomy ID, passed in the call to [get_counts()]).
}
\description{
Returns a summary data frame containing relevant counts for each taxID
associated with the sequences passed to [get_counts()]
(parameters `ids_df` or `ids_file`) in the pre-processing stage before the
call to [run_TS()].
}
