% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxID_spp_counts.R
\name{get_taxID_spp_counts}
\alias{get_taxID_spp_counts}
\title{Count number of species under each taxonomy ID}
\usage{
get_taxID_spp_counts(
  taxonomy_path = NULL,
  nodes = NULL,
  spp_file = NULL,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{taxonomy_path}{path to folder containing the NCBI taxonomy files
(i.e., the extracted contents of _taxdump.zip_, which can be downloaded from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/> or retrieved using
[retrieve_NCBI_taxonomy()]).}

\item{nodes}{data frame of taxonomic node information, containing at least
columns `$id` (taxonomy ID of the row), `$parent` (taxonomy ID of the
parent node of `$id`), and `$level` (taxonomic level of `$id`). A `nodes`
data frame of this format is part of objects of class _taxonsampling_,
generated by [get_taxonomy_counts()], [get_species_counts()] or
[run_TS()]. It can also be generated by [CHNOSZ::getnodes()].
Ignored if `taxonomy_path` is not `NULL`.}

\item{spp_file}{path to file for saving the species count for each taxon ID
(saved as a tsv file)}

\item{ncpus}{number of cores to use for species counting.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
data frame with species count per taxon ID in `nodes$id`.
}
\description{
Get a nodes data frame (generated, e.g., by [get_taxonomy_counts()] or
[CHNOSZ::getnodes()]) and returns the count of species under each taxonomy
ID.
}
