% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy_counts.R
\name{get_taxonomy_counts}
\alias{get_taxonomy_counts}
\title{Return the count of each taxon in a taxonomy graph}
\usage{
get_taxonomy_counts(
  taxonomy_path,
  ids_file = NULL,
  ids_df = NULL,
  verbose = TRUE
)
}
\arguments{
\item{taxonomy_path}{path to folder containing the NCBI taxonomy files
(i.e., the contents of _taxdump.zip_, which can be downloaded from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/> or retrieved using
[retrieve_NCBI_taxonomy()]).}

\item{ids_file}{path to a tab-separated file containng two two columns,
with the input taxon IDs in the first
column, and the corresponding sequence IDs (corresponding
to the ID strings int the multifasta input file, without
the ">" line starter) in the second column.}

\item{ids_df}{two-column data frame with the input taxon IDs in the first
column, and the corresponding sequence IDs (corresponding
to the ID strings int the multifasta input file, without
the ">" line starter) in the second column. Ignored if
`ids_file` is not `NULL`.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
list object containing:
\itemize{
    \item `$nodes`: data.frame containing the pre-processed information about
    the NCBI taxonomy structure. This is extracted internally by calling
    [CHNOSZ::getnodes()].
    \item `$countIDs`: numeric vector with the counts of the number of
    taxonomy IDs belonging to each taxon.
}
}
\description{
This function evaluates the taxonomy structure from a set of input IDs and
returns the count of each taxon ID with non-zero occurrences in the
taxonomy graph.
}
