% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxonsampling.R
\name{plot.taxonsampling}
\alias{plot.taxonsampling}
\title{plot.taxonsampling}
\usage{
\method{plot}{taxonsampling}(x, ...)
}
\arguments{
\item{x}{list object of class _taxonsampling_
(generated by [run_TS()])}

\item{...}{other parameters to be passed down to specific
plot functions (currently unused)}
}
\value{
Nothing (called for side effects)
}
\description{
S3 method for plotting _taxonsampling_ objects output by [run_TS()]).
}
