% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TS.R
\name{run_TS}
\alias{run_TS}
\title{Run Taxon Sampling}
\usage{
run_TS(
  taxlist,
  taxon,
  m,
  seq_file = NULL,
  out_file = NULL,
  method = "diversity",
  randomize = "no",
  replacement = FALSE,
  ignoreIDs = NULL,
  requireIDs = NULL,
  sampling = "agnostic",
  verbose = TRUE
)
}
\arguments{
\item{taxlist}{list object of class _taxonsampling_, returned by
[get_species_count()].}

\item{taxon}{Taxon ID from which to start sampling children taxa (single
character or integer value)}

\item{m}{desired sample size}

\item{seq_file}{character string with the path to the multifasta file
containing the input sequences.}

\item{out_file}{character string naming a file to save the output (a
multifasta file). Ignored if `seq_file == NULL`.}

\item{method}{sampling method to use. Accepts "balance" (favors balanced taxa
representation) or "diversity" (favors maximized taxa representation)}

\item{randomize}{randomization strategy: should the algorithm choose IDs
randomly ("yes"), maintaining a balanced allocation ("no"), or with a
balanced allocation at the top taxonomic level and randomized afterwards
("after_first_round")?}

\item{replacement}{logical flag: should the algorithm allow repeated IDs in
the output (if needed to reach m IDs in the output with maximized taxonomy
diversity).}

\item{ignoreIDs}{vector (character or integer) of IDs that must not appear in
the output.}

\item{requireIDs}{vector (character or integer) of IDs that must appear
in the output. Notice that `ignoreIDs` has precedence over `requireIDs`,
i.e., IDs that occur in both will be ignored. `requireIDs` that are children
of any `ignoreIDs` will also be ignored.}

\item{sampling}{sampling mode. Accepts "agnostic" (sample species in a
diversity-agnostic manner) or "known_species" (sample based on known species
diversity).}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
Input object`taxlist` updated with vector `$outputIDs` of
sampled IDs and list `$outputSeqs` (if seq_file is not `NULL`) containing
information about the sequences sampled.
}
\description{
Run the TaxonSampling method to return a sample of taxonomic IDs according
to the desired balance / diversity.
}
