% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_taxonsampling.R
\name{summary.taxonsampling}
\alias{summary.taxonsampling}
\title{summary.taxonsampling}
\usage{
\method{summary}{taxonsampling}(object, ...)
}
\arguments{
\item{object}{list object of class _taxonsampling_
(generated by [run_TS()])}

\item{...}{other parameters to be passed down to specific
summary functions (currently unused)}
}
\value{
Nothing (called for side effects)
}
\description{
S3 method for summarizing _taxonsampling_ objects output by [run_TS()]).
}
