% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_counts.R
\name{get_species_counts}
\alias{get_species_counts}
\title{Return the count of known species in each taxon}
\usage{
get_species_counts(taxlist, spp_df = NULL, spp_file = NULL, verbose = TRUE)
}
\arguments{
\item{taxlist}{a list object of class _taxonsampling_, returned by
[get_taxonomy_counts()]).}

\item{spp_df}{two-column data frame with the input taxon IDs in the first
column, and the corresponding number of known species in the second column.
Ignored if `ids_file` is not `NULL`.}

\item{spp_file}{path to a tab-separated file containng two two columns,
with the input taxon IDs in the first
column, and the corresponding number of known species in the second column.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
Input object `taxlist` updated to contain the additional fields:
\itemize{
   \item `spp_df`: two-column data frame with the input taxon IDs in the
   first column, and the corresponding number of known species in the second
   column. Filtered to have only the IDs present in `$countIDs`.
   \item `$countSpp`, a numeric vector with the counts of (known) species for each
taxon from `taxlist$countIDs`.
}
}
\description{
This function processes a list of NCBI Taxon IDs and known number of species
and determines the count of species under each taxon
}
