% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxID_spp_counts.R
\name{get_taxID_spp_counts}
\alias{get_taxID_spp_counts}
\title{Count number of species under each taxonomy ID}
\usage{
get_taxID_spp_counts(
  taxonomy_path = NULL,
  nodes = NULL,
  out_file = NULL,
  start_from_species = FALSE,
  what = "spp_counts",
  verbose = TRUE
)
}
\arguments{
\item{taxonomy_path}{path to folder containing the NCBI taxonomy files
(i.e., the extracted contents of _taxdump.zip_, which can be downloaded from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/> or retrieved using
[retrieve_NCBI_taxonomy()]). Ignored if `nodes` is not `NULL`.}

\item{nodes}{data.frame containing the pre-processed information about
the NCBI taxonomy structure. This can be generated, e.g., by using
[CHNOSZ::getnodes()].}

\item{out_file}{path to file for saving the output
(saved as a tsv file). Ignored if `NULL`.}

\item{start_from_species}{logical. If `TRUE` the counting starts at
the 'species' level - i.e., taxon IDs below species (e.g., 'subspecies')
receive a count of zero and are not included in the other counts. If
`FALSE`, counting starts at the terminal nodes of the taxonomy, regardless of
taxonomic level).}

\item{what}{what to calculate? Accepts `"spp_counts"` for returning the
count of species / leaf nodes per taxon ID, or `"ancestry"` for returning
the ancestry list.}

\item{verbose}{logical: regulates function echoing to console.}
}
\value{
either a data frame with species counts per taxon ID
or a list with the ancestry of each taxonomic ID.
}
\description{
Extract the species count for each taxonomy ID in the _nodes_ structure
of the NCBI taxonomy.
}
