% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxID_spp_counts.R
\name{get_taxID_spp_counts}
\alias{get_taxID_spp_counts}
\title{Count number of species under each taxonomy ID}
\usage{
get_taxID_spp_counts(taxonomy_path, spp_file, ncpus = 1)
}
\arguments{
\item{taxonomy_path}{path to folder containing the NCBI taxonomy files
(i.e., the extracted contents of _taxdump.zip_, which can be downloaded from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/> or retrieved using
[retrieve_NCBI_taxonomy()]).}

\item{spp_file}{path to file for saving the species count for each taxon ID
(saved as a tsv file)}

\item{ncpus}{number of cores to use for species counting.}
}
\value{
data frame with species counts per taxon ID.
}
\description{
Extract the species count for each taxonomy ID in the _nodes_ structure
of the NCBI taxonomy.
}
