% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_NCBI_taxonomy.R
\name{retrieve_NCBI_taxonomy}
\alias{retrieve_NCBI_taxonomy}
\title{Retrieve NCBI taxonomy file from public database}
\usage{
retrieve_NCBI_taxonomy(
  target.dir,
  method = "auto",
  unzip = getOption("unzip"),
  url = "https://ftp.ncbi.nih.gov/pub/taxonomy/new_taxdump/new_taxdump.zip",
  timeout = 600,
  ...
)
}
\arguments{
\item{target.dir}{path to the folder where the files will be saved (
accepts relative and absolute paths)}

\item{method}{Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl", "wget" and
"curl", and there is a value "auto": see _Details_ and _Note_ in the
documentation of \code{utils::download.file()}.}

\item{unzip}{The unzip method to be used. See the documentation of
\code{utils::unzip()} for details.}

\item{url}{URL of the full NCBI taxonomy file.}

\item{timeout}{maximum time allowed for the download, in seconds.
Increase when under a slow connection.}

\item{...}{additional attributes (currently ignored)}
}
\value{
No return value, called for side effects (see Description).
}
\description{
This script downloads the full NCBI taxonomy from
<ftp://ftp.ncbi.nih.gov/pub/taxonomy/new_taxdump/new_taxdump.tar.gz> and
extracts it to the folder in `target.dir`.
}
\examples{
\dontrun{
  TaxonSampling::retrieve_NCBI_taxonomy(target.dir = "data_files/taxdump")
}

}
